requirements.txt
Django>=4.2,<6.1
graphene-django>=3.0.0
djangorestframework-simplejwt>=5.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
dj-database-url>=2.1.0
